+++ dubbo-serialization-extensions/dubbo-serialization-fory/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.common.serialize.Serialization
@@ -1,0 +1,2 @@
+fory=org.apache.dubbo.common.serialize.fory.dubbo.ForySerialization
+fory-compatible=org.apache.dubbo.common.serialize.fory.dubbo.ForyCompatibleSerialization
