+++ dubbo-serialization-extensions/dubbo-serialization-fory/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.model.ScopeModelInitializer
@@ -1,0 +1,2 @@
+fory=org.apache.dubbo.common.serialize.fory.dubbo.ForyScopeModelInitializer
+fory-compatible=org.apache.dubbo.common.serialize.fory.dubbo.ForyScopeModelInitializer
