--- dubbo-serialization-extensions/dubbo-serialization-fury/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.common.serialize.Serialization
@@ -1,2 +1,0 @@
-fury=org.apache.dubbo.common.serialize.fury.dubbo.FurySerialization
-fury-compatible=org.apache.dubbo.common.serialize.fury.dubbo.FuryCompatibleSerialization
