--- dubbo-serialization-extensions/dubbo-serialization-fury/src/main/resources/META-INF/dubbo/internal/org.apache.dubbo.rpc.model.ScopeModelInitializer
@@ -1,2 +1,0 @@
-fury=org.apache.dubbo.common.serialize.fury.dubbo.FuryScopeModelInitializer
-fury-compatible=org.apache.dubbo.common.serialize.fury.dubbo.FuryScopeModelInitializer
