+++ dubbo-serialization-extensions/pom.xml
@@ -39,7 +39,7 @@
         <module>dubbo-serialization-gson</module>
         <module>dubbo-serialization-fst</module>
         <module>dubbo-serialization-fastjson</module>
-        <module>dubbo-serialization-fury</module>
+        <module>dubbo-serialization-fory</module>
         <module>dubbo-serialization-avro</module>
         <module>dubbo-serialization-msgpack</module>
         <module>dubbo-serialization-native-hessian</module>
